// frontend/src/types/bookingRequest.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { BookingRequestType } from './bookingRequestType';
import type { BookingRequestStatus } from './bookingRequestStatus';

export interface BookingRequest {
  /** @minLength 1 */
  name: string;
  email?: string;
  /** @minLength 1 */
  phone: string;
  type: BookingRequestType;
  note?: string;
  date: string;
  time: string;
  status: BookingRequestStatus;
}